oba_AI/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
openai==1.44.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.0
pymongo==4.7.2
